Add files via upload
This is a fixed version of the OSSEC WUI that has most of the legacy curly brackets converted to square brackets. It makes most of the WUI functional. This zip is meant to be used my video here on how to deploy OSSEC + OSSEC WUI. 

ALWAYS inspect foreign code before deploying in a production environment. Do not blindly trust.